Version 1
The .csv you are using needs to be set in line 6. It should be formatted to have the required variables in line 41.

Todo:
document variable formatting in csv file
test use of try/catch in code for variable passthrough diagnostics
add filecheck for variables, and to notify if NA or NULL or ERROR.
CreateNew OR Attach VHD - Path and Size
configure each?